requirements.txt
pandas
numpy
python-dotenv
tiingo
sqlalchemy
matplotlib
seaborn